[mpd-show] Fixed bug in song formatting and added %Y% pseudo tag
Fixed minor bug in song formatting which caused groups to be skipped
incorrectly.  Also added %Y% tag which expands into an empty string but
is treated as if it produced non-empty string.

Also, a minor change was the non-hilighted colour was changed from
bold-white to default colour as it was in old version of mpd-show.