[cdiff] Better handling of recursive diff
When diff is called with -r switch the first line of output is
a diff command.  It is enought to detect diff mode but would
require much more parsing.  Instead, if the first line does not
match any patters used for detecting modes next line is
considered and so on.

Also, the code has been refactored by splitting it into several
more functions (compiler should inline them anyways).