[cdiff] Fix bug where long lines were not colourised
This commit fixes a bug where cdiff would not colourise long
lines (longer then 4095 characters) and instead treat each
4096-char long strip as a separate line.

This bug was caused by not checking whether fgets(3) read the
whole line (that is whether data read by fgets(3) ends with
a new line) and instead assume it had.

In practise this meant that only the first 4096 characters of
line were colourised.  In theory this could mean that the rest
of the line could be colourised differently.