[mpd-show] Silence gcc unused result warninng
When compiling mpd-show, gcc started complaining:

	mpd-show.c:1131: warning: ignoring return value of 'wctomb',
		declared with attribute warn_unused_result

which was not really applicable in given location since wctomb(NULL, 0)
returns only a piece of information that can be easily ignored.  Only
wctomb(3) invocations with non-NULL argument return an error code.

The warning is silenced by assigning the return value to a variable,
which was there already, but otherwise ignoring it.