[null] Remove some of the #if directives and compiler warnings
This commit removes some of the #if directives instead relaying on the
compiler that it will remove “if (0) ...” portions of the code.

It also provides a dummy definition of WCOREDUMP() macro (to return
0) so that a #if can be avoided in the code.

Lastly, it adds dummy if statements around nice() and chdir() since
apparently, at one point, glib added a warn_unused_result attribute to
them.

Lastly, it introduces some minor style changes and documentation
fixes.