[mpd-show] Fix string comparison in string formatting
The EQ() macro in doFormatTag() compared only the first strlen(str)
bytes even though it was comparing wchar_t strings.

The end effect was that only the length and the first half of the
characters had to match for tag to be recognised.  Fortunately, this
did not cause any valid formats to be unrecognised.

This commit multiples number of characters by sizeof(wchar_t) and
therefore makes the memcmp() compare all of the bytes in the wchar_t
sub-string.