[genpass] Improved “strength” handling.
The “strength” option's purpose is to ensure that the password has
at least one upper case letter, one lower case letter and one
digit, which is a required by some websites and with a truly
random passwords it is possible to end up with one with, say, no
digits.

Previously the code just added “1aA” to the end of the password,
which really wasted opportunity for some more randomness.  With
this commit, the code now adds random characters plus if password
so far already has character of given class, a random char is added
without the restriction.

This ends up adding another 12-19 bits of randomness (depends on set
of characters chosen).