Migrate away from Academic Free License v. 3.0.
This commit changes the license of most of the software that used
Academic Free License v. 3.0 to Apache License v. 2.0.  Two
exceptions are: the show script whom I'm not the sole copyright
holder, and the quetes.txt file which is now licensed under
Creative Commons Attribution 3.0 Unported license which is more
appropriate for non-software files.

It also updates some copyright lines with current date.