[gz2bz] Added support for keeping smaller of the files
The main use case of the gz2bz script is to recompress files
with an algorithm that gives higher compression.  However,
in some cases (espacially small files), a “worse” algorithm
may yield better results (due to overhead I presume).  In
those cases, it is better to keep the original file rather
then the recompressed one.

This commit adds --keep-smaller (also -s or --smaller)
switch will causes the script to compare sizes of the
original and recompressed files and delete the bigger one.
With --keep-original also given, the original file is
never deleted, but the recompressed one might if it is
bigger then the original.