[surun] New tool for running commands asa root
It uses sudo or su (whichever is available) to actually become root,
but thanks to automatic detection hides some differences in the way
those tools are called.

su is particularly annoying if one wants to run a command as
super-user since it requires the whole command to be escaped and
specified as shell expression.  sudo on the other hand requires
“-s” argument if one wants to run a shell.

With surun, one spawns a root shell by calling it with no arguments
(like su), and runs a command by specifying the command as arguments
(like sudo).