[tpwd] Get rid of specialised code and prefer sourcing to evaluation
Instead of having three separate versions of the function – one executed
when the script is run, one for bash and zsh and another for POSIX
shells – use a single code that works in all shells.  This simplifies
the code – since now there's only one version to maintain – and allows
the file to be sourced rather than having to eval its output when called
with --function or so.  The --function, --bash and --zsh options are now
deprecated.