[tpwd] Add support for TPWD_DIRS environment variable
TPWD_DIRS may be set to a colon-separated list of environment variable
names.

If one of the variable is set to an absolute path (i.e. its value starts
with a slash) which is a path prefix of current working directory, the
printed value is shortened by replacing this prefix by ‘$VARNAME’
string.

If TPWD_DIRS variable is set the usual ‘~’ shortening is not performed.
To explicitly request it, ‘~’ must be given as one of the entry of the
TPWD_DIRS variable.  I.e. TPWD_DIRS=~ is identical with the default
behaviour.

Variables are checked in the order they are given in the TPWD_DIRS
variable and substitution happens only if it resurts in a shorter
string.  For example:

    $ cd /etc
    $ CONFIG_DIR=/etc TPWD_DIRS='CONFIG_DIR:~' tpwd
    /etc