mpd-show: initialise iconv before using it
Since parseArguments was called before initCodesets it could not use
iconv for converting multibyte characters into wide characters but
that’s exactly what that function attempted to do with the help of
wideFromMulti function.  Because of this the fall-back code had to
be used instead of using iconv.  Swap the initialisation order so
that iconv is set up first.

Also, fix the fall-back code in wideFromMulti where it incorrectly
handled mbrtowc returning -2.